requirements.txt
Flask
pdfplumber
python-docx
pytesseract
Pillow
spacy
matplotlib
gunicorn